Refactor python code generator to support other serializers
