+++ build.yaml
@@ -1042,7 +1042,6 @@ libs:
   headers:
   - src/compiler/config.h
   - src/compiler/cpp_generator.h
-  - src/compiler/cpp_generator_helpers.h
   - src/compiler/csharp_generator.h
   - src/compiler/csharp_generator_helpers.h
   - src/compiler/generator_helpers.h
@@ -1051,6 +1050,7 @@ libs:
   - src/compiler/objective_c_generator.h
   - src/compiler/objective_c_generator_helpers.h
   - src/compiler/python_generator.h
+  - src/compiler/python_generator_helpers.h
   - src/compiler/ruby_generator.h
   - src/compiler/ruby_generator_helpers-inl.h
   - src/compiler/ruby_generator_map-inl.h
