+++ tools/run_tests/sources_and_headers.json
@@ -4484,7 +4484,6 @@
     "headers": [
       "src/compiler/config.h", 
       "src/compiler/cpp_generator.h", 
-      "src/compiler/cpp_generator_helpers.h", 
       "src/compiler/csharp_generator.h", 
       "src/compiler/csharp_generator_helpers.h", 
       "src/compiler/generator_helpers.h", 
@@ -4492,7 +4491,8 @@
       "src/compiler/node_generator_helpers.h", 
       "src/compiler/objective_c_generator.h", 
       "src/compiler/objective_c_generator_helpers.h", 
-      "src/compiler/python_generator.h", 
+      "src/compiler/python_generator.h",
+      "src/compiler/python_generator_helpers.h", 
       "src/compiler/ruby_generator.h", 
       "src/compiler/ruby_generator_helpers-inl.h", 
       "src/compiler/ruby_generator_map-inl.h", 
@@ -4503,8 +4503,7 @@
     "src": [
       "src/compiler/config.h", 
       "src/compiler/cpp_generator.cc", 
-      "src/compiler/cpp_generator.h", 
-      "src/compiler/cpp_generator_helpers.h", 
+      "src/compiler/cpp_generator.h",
       "src/compiler/csharp_generator.cc", 
       "src/compiler/csharp_generator.h", 
       "src/compiler/csharp_generator_helpers.h", 
@@ -4516,7 +4515,8 @@
       "src/compiler/objective_c_generator.h", 
       "src/compiler/objective_c_generator_helpers.h", 
       "src/compiler/python_generator.cc", 
-      "src/compiler/python_generator.h", 
+      "src/compiler/python_generator.h",
+      "src/compiler/python_generator_helpers.h", 
       "src/compiler/ruby_generator.cc", 
       "src/compiler/ruby_generator.h", 
       "src/compiler/ruby_generator_helpers-inl.h",
